feat: add two tests for the logging functions #40
