+++ ci_server/server.py
@@ -4,6 +4,7 @@ from http.server import HTTPServer
 from dotenv import load_dotenv
 import requests
 from .handler import CIServerHandler
+import time
 
 
 class CIServer:
@@ -48,3 +49,36 @@ class CIServer:
         DATA = {"state": statusString}
         response = requests.post(url=URL, data=json.dumps(DATA), headers=HEADERS)
         print(response)
+
+        # LOGGING
+
+    # The purpose of this function is to generate a unique identifier that serves
+    # as the build log name.
+    # Note: This would fail if the number of total builds would reach maxint.
+    def make_log_title(self):
+        tob = time.localtime()
+        bString = "X"
+        bDatPath = "logfiles/buildData.dat"
+        newBuildNum = -1
+
+        # Format the new build number.
+        with open(bDatPath) as f:
+            newBuildNum = int(f.read()[:-1]) + 1
+
+        # Generate a build log string.
+        bString = f"Build_{newBuildNum}_{tob.tm_year}_{tob.tm_mon}_{tob.tm_mday}_{tob.tm_hour}.txt"
+
+        with open(bDatPath, "w") as f:
+            f.truncate(0)
+            f.write(str(newBuildNum) + "\n")
+
+        return bString
+
+    # The purpose of this function is to log the output of the tests and the linting.
+    # Creates a new .txt file with the output of the linter and the tests.
+    def make_log(self, lint_output, pytest_output):
+        with open(f"logfiles/{self.make_log_title()}", "w") as f:
+            f.write(
+                f"=== LINT OUTPUT ===\n{lint_output}\n\n=== PYTEST OUTPUT ===\n{pytest_output}\n"
+            )
+
